requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
motor==3.6.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
streamlit
requests
